backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0

# LLM providers (install as needed)
openai>=1.10.0
anthropic>=0.18.0

# Development
python-dotenv>=1.0.0
